ds info log on rank0 only
